requirements.txt
streamlit
pandas
plotly
vaderSentiment
wordcloud
matplotlib
numpy
Pillow
